refactor: move search page to app router
