+++ src/app/(core)/teachers/search/page.test.tsx
@@ -1,10 +1,10 @@
 import { usePathname } from "next/navigation";
 
-import SearchPage from "@/pages/teachers/search";
+import { SearchView } from "./SearchView";
+
 import { mockSeoResult } from "@/__tests__/__helpers__/cms";
 import renderWithSeo from "@/__tests__/__helpers__/renderWithSeo";
 import searchPageFixture from "@/node-lib/curriculum-api-2023/fixtures/searchPage.fixture";
-import { topNavFixture } from "@/node-lib/curriculum-api-2023/fixtures/topNav.fixture";
 
 const fixture = searchPageFixture();
 if (!fixture) {
@@ -27,8 +27,7 @@ jest.mock("posthog-js/react", () => ({
 describe("pages/teachers/search.tsx", () => {
   test("renders page with correct seo", () => {
     const { seo } = renderWithSeo()(
-      <SearchPage
-        topNav={topNavFixture}
+      <SearchView
         curriculumData={{
           keyStages,
           subjects,
@@ -54,8 +53,7 @@ describe("pages/teachers/search.tsx", () => {
 
   test("renders correct content type filters", () => {
     const { getAllByRole } = renderWithSeo()(
-      <SearchPage
-        topNav={topNavFixture}
+      <SearchView
         curriculumData={{
           keyStages,
           subjects,
@@ -78,8 +76,7 @@ describe("pages/teachers/search.tsx", () => {
 
   test("renders correct new curriculum filter", async () => {
     const { getAllByRole } = renderWithSeo()(
-      <SearchPage
-        topNav={topNavFixture}
+      <SearchView
         curriculumData={{
           keyStages,
           subjects,
@@ -97,8 +94,7 @@ describe("pages/teachers/search.tsx", () => {
 
   test("renders correct key stage filters", async () => {
     const { getAllByRole } = renderWithSeo()(
-      <SearchPage
-        topNav={topNavFixture}
+      <SearchView
         curriculumData={{
           keyStages,
           subjects,
@@ -117,8 +113,7 @@ describe("pages/teachers/search.tsx", () => {
 
   test("renders correct year group filters", () => {
     const { getAllByRole } = renderWithSeo()(
-      <SearchPage
-        topNav={topNavFixture}
+      <SearchView
         curriculumData={{
           keyStages,
           subjects,
@@ -136,8 +131,7 @@ describe("pages/teachers/search.tsx", () => {
 
   test("renders correct examBoard filters", async () => {
     const { getAllByRole } = renderWithSeo()(
-      <SearchPage
-        topNav={topNavFixture}
+      <SearchView
         curriculumData={{
           keyStages,
           subjects,
@@ -156,8 +150,7 @@ describe("pages/teachers/search.tsx", () => {
 
   test("renders correct subject filters", () => {
     const { getAllByRole } = renderWithSeo()(
-      <SearchPage
-        topNav={topNavFixture}
+      <SearchView
         curriculumData={{
           keyStages,
           subjects,
