+++ src/app/(core)/teachers/search/page.tsx
@@ -1,0 +1,23 @@
+import React from "react";
+import { Metadata } from "next";
+
+import { SearchView } from "./SearchView";
+
+import curriculumApi2023 from "@/node-lib/curriculum-api-2023";
+
+export const metadata: Metadata = {
+  title: "Search for Free Teaching Resources",
+  description: "Search for Free Teaching Resources",
+  robots: {
+    index: false,
+    follow: true,
+  },
+};
+
+async function SearchPage() {
+  const curriculumData = await curriculumApi2023.searchPage();
+
+  return <SearchView curriculumData={curriculumData} />;
+}
+
+export default SearchPage;
